Export types instead of enums #418
Export what were previously enums as types

Update the typings of `Constants` so they are more type-safe.

Resolves #418
