+++ index.d.ts
@@ -1,16 +1,16 @@
+import {
+  HealthActivity,
+  HealthObserver,
+  HealthPermission,
+  HealthUnit,
+} from './enums'
+
 declare module 'react-native-health' {
   export interface HealthKitPermissions {
     permissions: {
       read: HealthPermission[]
       write: HealthPermission[]
     }
-  }
-
-  export interface Constants {
-    Activities: Record<HealthActivity, HealthActivity>
-    Observers: Record<HealthObserver, HealthObserver>
-    Permissions: Record<HealthPermission, HealthPermission>
-    Units: Record<HealthUnit, HealthUnit>
   }
 
   export interface HKErrorResponse {
@@ -462,8 +462,12 @@ declare module 'react-native-health' {
       callback: (error: string, result: HealthValue) => void,
     ): void
 
-
-    Constants: Constants
+    Constants: {
+      Activities: typeof HealthActivity
+      Observers: typeof HealthObserver
+      Permissions: typeof HealthPermission
+      Units: typeof HealthUnit
+    }
   }
 
   /* Inputs and Payloads */
@@ -546,7 +550,7 @@ declare module 'react-native-health' {
     PausedOrResumeRequest = 'pause or resume request',
     Lap = 'lap',
     Segment = 'segment',
-    Marker = 'marker'
+    Marker = 'marker',
   }
 
   export type HKWorkoutEventType = {
@@ -633,211 +637,10 @@ declare module 'react-native-health' {
     fhirData: any
   }
 
-  /* Health Constants */
-
-  export enum HealthActivity {
-    AmericanFootball = 'AmericanFootball',
-    Archery = 'Archery',
-    AustralianFootball = 'AustralianFootball',
-    Badminton = 'Badminton',
-    Baseball = 'Baseball',
-    Basketball = 'Basketball',
-    Bowling = 'Bowling',
-    Boxing = 'Boxing',
-    CardioDance = 'CardioDance',
-    Climbing = 'Climbing',
-    Cooldown = 'Cooldown',
-    Cricket = 'Cricket',
-    CrossTraining = 'CrossTraining',
-    Curling = 'Curling',
-    Cycling = 'Cycling',
-    Dance = 'Dance',
-    DiscSports = 'DiscSports',
-    Elliptical = 'Elliptical',
-    EquestrianSports = 'EquestrianSports',
-    Fencing = 'Fencing',
-    FitnessGaming = 'FitnessGaming',
-    Fishing = 'Fishing',
-    FunctionalStrengthTraining = 'FunctionalStrengthTraining',
-    Golf = 'Golf',
-    Gymnastics = 'Gymnastics',
-    Handball = 'Handball',
-    Hiking = 'Hiking',
-    Hockey = 'Hockey',
-    Hunting = 'Hunting',
-    Lacrosse = 'Lacrosse',
-    MartialArts = 'MartialArts',
-    MindAndBody = 'MindAndBody',
-    PaddleSports = 'PaddleSports',
-    Play = 'Play',
-    Pickleball = 'Pickleball',
-    PreparationAndRecovery = 'PreparationAndRecovery',
-    Racquetball = 'Racquetball',
-    Rowing = 'Rowing',
-    Rugby = 'Rugby',
-    Running = 'Running',
-    Sailing = 'Sailing',
-    SkatingSports = 'SkatingSports',
-    SnowSports = 'SnowSports',
-    Soccer = 'Soccer',
-    SocialDance = 'SocialDance',
-    Softball = 'Softball',
-    Squash = 'Squash',
-    StairClimbing = 'StairClimbing',
-    SurfingSports = 'SurfingSports',
-    Swimming = 'Swimming',
-    TableTennis = 'TableTennis',
-    Tennis = 'Tennis',
-    TrackAndField = 'TrackAndField',
-    TraditionalStrengthTraining = 'TraditionalStrengthTraining',
-    Volleyball = 'Volleyball',
-    Walking = 'Walking',
-    WaterFitness = 'WaterFitness',
-    WaterPolo = 'WaterPolo',
-    WaterSports = 'WaterSports',
-    Wrestling = 'Wrestling',
-    Yoga = 'Yoga',
-    Barre = 'Barre',
-    CoreTraining = 'CoreTraining',
-    CrossCountrySkiing = 'CrossCountrySkiing',
-    DownhillSkiing = 'DownhillSkiing',
-    Flexibility = 'Flexibility',
-    HighIntensityIntervalTraining = 'HighIntensityIntervalTraining',
-    JumpRope = 'JumpRope',
-    Kickboxing = 'Kickboxing',
-    Pilates = 'Pilates',
-    Snowboarding = 'Snowboarding',
-    Stairs = 'Stairs',
-    StepTraining = 'StepTraining',
-    WheelchairWalkPace = 'WheelchairWalkPace',
-    WheelchairRunPace = 'WheelchairRunPace',
-    TaiChi = 'TaiChi',
-    MixedCardio = 'MixedCardio',
-    HandCycling = 'HandCycling',
-  }
-
-  export enum HealthPermission {
-    ActiveEnergyBurned = 'ActiveEnergyBurned',
-    ActivitySummary = 'ActivitySummary',
-    AllergyRecord = 'AllergyRecord',
-    AppleExerciseTime = 'AppleExerciseTime',
-    AppleStandTime = 'AppleStandTime',
-    BasalEnergyBurned = 'BasalEnergyBurned',
-    BiologicalSex = 'BiologicalSex',
-    BloodType = 'BloodType',
-    BloodAlcoholContent = 'BloodAlcoholContent',
-    BloodGlucose = 'BloodGlucose',
-    BloodPressureDiastolic = 'BloodPressureDiastolic',
-    BloodPressureSystolic = 'BloodPressureSystolic',
-    BodyFatPercentage = 'BodyFatPercentage',
-    BodyMass = 'BodyMass',
-    BodyMassIndex = 'BodyMassIndex',
-    BodyTemperature = 'BodyTemperature',
-    DateOfBirth = 'DateOfBirth',
-    Biotin = 'Biotin',
-    Caffeine = 'Caffeine',
-    Calcium = 'Calcium',
-    Carbohydrates = 'Carbohydrates',
-    Chloride = 'Chloride',
-    Cholesterol = 'Cholesterol',
-    ConditionRecord = 'ConditionRecord',
-    Copper = 'Copper',
-    CoverageRecord = 'CoverageRecord',
-    EnergyConsumed = 'EnergyConsumed',
-    EnvironmentalAudioExposure = 'EnvironmentalAudioExposure',
-    FatMonounsaturated = 'FatMonounsaturated',
-    FatPolyunsaturated = 'FatPolyunsaturated',
-    FatSaturated = 'FatSaturated',
-    FatTotal = 'FatTotal',
-    Fiber = 'Fiber',
-    Folate = 'Folate',
-    HeadphoneAudioExposure = 'HeadphoneAudioExposure',
-    ImmunizationRecord = 'ImmunizationRecord',
-    InsulinDelivery = 'InsulinDelivery',
-    Iodine = 'Iodine',
-    Iron = 'Iron',
-    LabResultRecord = 'LabResultRecord',
-    Magnesium = 'Magnesium',
-    Manganese = 'Manganese',
-    MedicationRecord = 'MedicationRecord',
-    Molybdenum = 'Molybdenum',
-    Niacin = 'Niacin',
-    OxygenSaturation = 'OxygenSaturation',
-    PantothenicAcid = 'PantothenicAcid',
-    Phosphorus = 'Phosphorus',
-    Potassium = 'Potassium',
-    ProcedureRecord = 'ProcedureRecord',
-    Protein = 'Protein',
-    Riboflavin = 'Riboflavin',
-    Selenium = 'Selenium',
-    Sodium = 'Sodium',
-    Sugar = 'Sugar',
-    Thiamin = 'Thiamin',
-    VitalSignRecord = 'VitalSignRecord',
-    VitaminA = 'VitaminA',
-    VitaminB12 = 'VitaminB12',
-    VitaminB6 = 'VitaminB6',
-    VitaminC = 'VitaminC',
-    VitaminD = 'VitaminD',
-    VitaminE = 'VitaminE',
-    VitaminK = 'VitaminK',
-    Zinc = 'Zinc',
-    Water = 'Water',
-    DistanceCycling = 'DistanceCycling',
-    DistanceSwimming = 'DistanceSwimming',
-    DistanceWalkingRunning = 'DistanceWalkingRunning',
-    Electrocardiogram = 'Electrocardiogram',
-    FlightsClimbed = 'FlightsClimbed',
-    HeartbeatSeries = 'HeartbeatSeries',
-    HeartRate = 'HeartRate',
-    RestingHeartRate = 'RestingHeartRate',
-    HeartRateVariability = 'HeartRateVariability',
-    Height = 'Height',
-    LeanBodyMass = 'LeanBodyMass',
-    MindfulSession = 'MindfulSession',
-    NikeFuel = 'NikeFuel',
-    PeakFlow = 'PeakFlow',
-    RespiratoryRate = 'RespiratoryRate',
-    SleepAnalysis = 'SleepAnalysis',
-    StepCount = 'StepCount',
-    Steps = 'Steps',
-    Vo2Max = 'Vo2Max',
-    WaistCircumference = 'WaistCircumference',
-    WalkingHeartRateAverage = 'WalkingHeartRateAverage',
-    Weight = 'Weight',
-    Workout = 'Workout',
-    WorkoutRoute = 'WorkoutRoute',
-    RunningSpeed = 'RunningSpeed',
-    RunningPower = 'RunningPower',
-    RunningStrideLength = 'RunningStrideLength',
-    RunningVerticalOscillation = 'RunningVerticalOscillation',
-    RunningGroundContactTime = 'RunningGroundContactTime',
-  }
-
-  export enum HealthUnit {
-    bpm = 'bpm',
-    calorie = 'calorie',
-    celsius = 'celsius',
-    count = 'count',
-    day = 'day',
-    fahrenheit = 'fahrenheit',
-    foot = 'foot',
-    gram = 'gram',
-    hour = 'hour',
-    inch = 'inch',
-    joule = 'joule',
-    kilocalorie = 'kilocalorie',
-    meter = 'meter',
-    mgPerdL = 'mgPerdL',
-    mile = 'mile',
-    minute = 'minute',
-    mmhg = 'mmhg',
-    literPerMinute = 'literPerMinute',
-    mmolPerL = 'mmolPerL',
-    percent = 'percent',
-    pound = 'pound',
-    second = 'second',
-  }
+  export type HealthActivity = keyof typeof HealthActivity
+  export type HealthPermission = keyof typeof HealthPermission
+  export type HealthUnit = keyof typeof HealthUnit
+  export type HealthObserver = keyof typeof HealthObserver
 
   export enum HealthStatusCode {
     NotDetermined = 0,
@@ -860,24 +663,6 @@ declare module 'react-native-health' {
   export interface WorkoutRouteQueryResults {
     anchor: string
     data: HKWorkoutRouteSampleType
-  }
-
-  export enum HealthObserver {
-    AllergyRecord = 'AllergyRecord',
-    ConditionRecord = 'ConditionRecord',
-    CoverageRecord = 'CoverageRecord',
-    Cycling = 'Cycling',
-    HeartRate = 'HeartRate',
-    ImmunizationRecord = 'ImmunizationRecord',
-    LabResultRecord = 'LabResultRecord',
-    MedicationRecord = 'MedicationRecord',
-    ProcedureRecord = 'ProcedureRecord',
-    RestingHeartRate = 'RestingHeartRate',
-    Running = 'Running',
-    StairClimbing = 'StairClimbing',
-    VitalSignRecord = 'VitalSignRecord',
-    Walking = 'Walking',
-    Workout = 'Workout',
   }
 
   export enum ElectrocardiogramClassification {
@@ -905,3 +690,10 @@ declare module 'react-native-health' {
 
   export default appleHealthKit
 }
+
+export const Constants = {
+  Activities: HealthActivity,
+  Observers: HealthObserver,
+  Permissions: HealthPermission,
+  Units: HealthUnit,
+} as const
